requirements.txt
google-auth==2.40.3
boto3==1.40.21
requests==2.32.5